requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
mypy-extensions==1.0.0
mysqlclient==2.2.7
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
PyJWT==2.9.0
sqlparse==0.5.3
